# Optimize `/solve` request serialization