# Description
Follow up to https://github.com/cowprotocol/services/pull/3386 since that actually did not resolve the issue.
We still have the problem where the first driver can see an auction sometimes multiple seconds before the last driver.
It turns out the problem is how we send out the requests. There are 2 main problems with our current approach:
1. the request is huge and we re-serialize it for each driver (this can take ~200ms)
2. we serialize the request in our main task

Both issues combined lead to the autopilot getting blocked for 200ms per driver. In the grand scheme of things this effectively turns our request sending from concurrent to serial. So if we have 20 solvers the last one could see the auction ~4s later than the first driver.

# Changes
I tackled `1` by introducing a new type that serializes the `/solve` request once and allows cheap clones with `Arc`. Serializing this type then effectively turns into copying a `str` instead of recursively calling the serialize logic of each contained struct.
`2` gets addresses by simply using `tokio::task::spawn_blocking()` when calling `request.json(data)`. Using the helper struct that's already way faster but just copying the data around can already take up to 30ms. So without this the 20th driver would still see the auction ~600ms later than the first one.

## How to test
I deployed the change to the mainnet shadow environment which for these purposes is identical to prod (huge requests, many connected drivers). With some temporary logs (not contained in this PR) I was able to verify that the autopilot now sends all requests within `~10ms` and all drivers received the requests within `~200ms`.

[Logs](https://aws-es.cow.fi/_dashboards/app/data-explorer/discover#?_a=(discover:(columns:!(log,log_level),isDirty:!f,sort:!()),metadata:(indexPattern:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',view:discover))&_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2025-05-15T13:24:29.054Z',to:'2025-05-15T13:35:24.105Z'))&_q=(filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',key:kubernetes.container_image,negate:!t,params:(query:nginx),type:phrase),query:(match_phrase:(kubernetes.container_image:nginx))),('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',key:kubernetes.pod_labels.network,negate:!f,params:(query:mainnet),type:phrase),query:(match_phrase:(kubernetes.pod_labels.network:mainnet)))),query:(language:kuery,query:'log:%20%2210673186%22%20and%20(log:%20%22solving%22%20or%20log:%20%22sending%20request%22%20or%20log:%20%22received%20auction%22)'))) without the patch (only shows difference between cutting the auction and drivers receiving it)
[Logs](https://aws-es.cow.fi/_dashboards/app/data-explorer/discover#?_a=(discover:(columns:!(log,log_level),isDirty:!f,sort:!()),metadata:(indexPattern:'5ed837d0-4a67-11ef-8b11-a98c7c46873d',view:discover))&_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2025-05-15T13:24:29.054Z',to:'2025-05-15T13:35:24.105Z'))&_q=(filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',key:kubernetes.container_image,negate:!t,params:(query:nginx),type:phrase),query:(match_phrase:(kubernetes.container_image:nginx))),('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',key:kubernetes.pod_labels.network,negate:!f,params:(query:mainnet),type:phrase),query:(match_phrase:(kubernetes.pod_labels.network:mainnet)))),query:(language:kuery,query:'%22shadow%22%20and%20log:%20%2210673180%22%20and%20(log:%20%22sending%20request%22%20or%20log:%20%22received%20auction%22)'))) with the patch (shows timings of sending and receiving the requests)